+++ redis/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3.0 (unreleased)
 
 * Add pipeline support
+* Add wrappers for `Cmd` and `Pipeline` to mimick the API of the `redis` crate
 
 ## v0.2.0